requirements.txt
google-generativeai
python-dotenv
Pillow
rembg
onnxruntime
streamlit
fastapi
uvicorn
python-multipart
